requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pydantic>=2.5.0
PyYAML>=6.0
typer[all]>=0.12.0
pydantic-settings>=2.1.0 